+++ migrations/2019_12_06_205537_create_tag_post_relation_table.php
@@ -16,9 +16,9 @@ class CreateTagPostRelationTable extends Migration
         if (!Schema::hasTable('tag_post_relation')) {
             Schema::create('tag_post_relation', function (Blueprint $table) {
                 $table->bigIncrements('id');
-                $table->unsignedInteger('user_id')->nullable(false)->comment('用户id');
-                $table->unsignedInteger('tag_id')->nullable(false)->comment('标签id');
-                $table->unsignedInteger('post_id')->nullable(false)->comment('帖子id');
+                $table->unsignedBigInteger('user_id')->nullable(false)->comment('用户id');
+                $table->unsignedBigInteger('tag_id')->nullable(false)->comment('标签id');
+                $table->unsignedBigInteger('post_id')->nullable(false)->comment('帖子id');
                 $table->unsignedBigInteger('created_at')->nullable();
                 $table->unsignedBigInteger('updated_at')->nullable();
             });
